Extract class methods from Proposal::SpaceDistribution
